refactor: reorg the various utils that we sadly have to have
