+++ iroh-bytes/src/util.rs
@@ -8,13 +8,15 @@ use range_collections::RangeSet2;
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 use std::{
     fmt::{self, Display},
-    io::{self, Read, Seek, Write},
+    io::{Read, Seek, Write},
     path::{Component, Path, PathBuf},
     result,
     str::FromStr,
 };
 use thiserror::Error;
 use tokio::sync::mpsc;
+pub mod io;
+pub mod progress;
 
 use crate::IROH_BLOCK_SIZE;
 
@@ -174,7 +176,7 @@ pub type RpcResult<T> = result::Result<T, RpcError>;
 #[derive(Debug, Display, Error)]
 pub enum BaoValidationError {
     /// Generic io error. We were unable to read the data.
-    IoError(#[from] io::Error),
+    IoError(#[from] std::io::Error),
     /// The data failed to validate
     EncodeError(#[from] EncodeError),
 }
@@ -204,7 +206,7 @@ pub fn validate_bao<F: Fn(u64)>(
 struct DevNull<F>(u64, F);
 
 impl<F: Fn(u64)> Write for DevNull<F> {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
+    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
         const NOTIFY_EVERY: u64 = 1024 * 1024;
         let prev = self.0;
         let curr = prev + buf.len() as u64;
@@ -215,7 +217,7 @@ impl<F: Fn(u64)> Write for DevNull<F> {
         Ok(buf.len())
     }
 
-    fn flush(&mut self) -> io::Result<()> {
+    fn flush(&mut self) -> std::io::Result<()> {
         Ok(())
     }
 }
