+++ iroh-bytes/src/util/io.rs
