+++ iroh/src/commands/doctor.rs
@@ -11,7 +11,7 @@ use crate::config::{iroh_data_root, Config};
 use anyhow::Context;
 use clap::Subcommand;
 use indicatif::{HumanBytes, MultiProgress, ProgressBar};
-use iroh_bytes::tokio_util::ProgressWriter;
+use iroh_bytes::util::io::ProgressWriter;
 use iroh_net::{
     defaults::DEFAULT_DERP_STUN_PORT,
     hp::{
